Quedo agregar, modificar y buscar propiedad

+++ project/src/main/java/com/javeriana/web/project/Properties/Property/Domain/ValueObjects/Latitude.java
@@ -3,24 +3,24 @@ package com.javeriana.web.project.Properties.Property.Domain.ValueObjects;
 import com.javeriana.web.project.Shared.Domain.InvalidNumber;
 
 public class Latitude {
-    private long value;
+    private float value;
 
     public Latitude(){
 
     }
 
-    public Latitude(long value) {
+    public Latitude(float value) {
         this.value = value;
         this.validate(value);
     }
 
-    private void validate(long value) {
+    private void validate(float value) {
         if(value<-90 || value>90) {
             throw new InvalidNumber("Invalid latitude number");
         }
     }
 
-    public long value(){
+    public float value(){
         return this.value;
     }
 }
